# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FormFlow API - AI Martial Arts Form Trainer
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthenticated

  # ─── Users ────────────────────────────────────────────────────────────────
...
  /dojos:
    get:
      operationId: listDojos
      tags: [dojos]
      summary: List all dojos (admin) or single dojo (dojo_admin)
      responses:
        "200":
          description: List of dojos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dojo"
    post:
      operationId: createDojo
      tags: [dojos]
      summary: Create a new dojo (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DojoInput"
      responses:
        "201":
          description: Created dojo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dojo"
  /dojos/{dojoId}:
    get:
      operationId: getDojo
      tags: [dojos]
      summary: Get dojo by ID
      parameters:
        - in: path
          name: dojoId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dojo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dojo"
    patch:
      operationId: updateDojo
      tags: [dojos]
      summary: Update dojo settings
      parameters:
        - in: path
          name: dojoId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DojoUpdate"
      responses:
        "200":
          description: Updated dojo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dojo"
    delete:
      operationId: deleteDojo
      tags: [dojos]
      summary: Delete a dojo (admin only)
      parameters:
        - in: path
          name: dojoId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /dojos/by-slug/{slug}:
    get:
      operationId: getDojoBySlug
      tags: [dojos]
      summary: Get public dojo info by slug (no auth required)
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public dojo info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DojoPublic"
  /dojos/{dojoId}/stats:
    get:
      operationId: getDojoStats
      tags: [dojos]
      summary: Get stats for a dojo (students, sessions, avg score)
      parameters:
        - in: path
          name: dojoId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dojo stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DojoStats"
  /dojos/{dojoId}/students:
    get:
      operationId: getDojoStudents
      tags: [dojos]
      summary: Get students in a dojo
      parameters:
        - in: path
          name: dojoId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
  /dojos/{dojoId}/invite:
    post:
      operationId: regenerateInviteCode
      tags: [dojos]
      summary: Regenerate the dojo invite code
      parameters:
        - in: path
          name: dojoId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: New invite code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteCode"
  /dojos/{dojoId}/leave:
    post:
      operationId: leaveDojo
      tags: [dojos]
      summary: Leave a dojo (student only)
      parameters:
        - in: path
          name: dojoId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
  /dojos/{dojoId}/remove-student/{userId}:
    delete:
      operationId: removeStudentFromDojo
      tags: [dojos]
      summary: Remove a student from dojo (dojo_admin only)
      parameters:
        - in: path
          name: dojoId
          required: true
          schema:
            type: string
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed
  /dojos/join:
    post:
      operationId: joinDojoByInvite
      tags: [dojos]
      summary: Join a dojo by invite code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinDojoInput"
      responses:
        "200":
          description: Joined dojo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dojo"

  # ─── Disciplines ──────────────────────────────────────────────────────────
  /disciplines:
    get:
      operationId: listDisciplines
      tags: [disciplines]
      summary: List all disciplines (public + dojo-specific)
      parameters:
        - in: query
          name: dojo_id
          schema:
            type: string
      responses:
        "200":
          description: List of disciplines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Discipline"
    post:
      operationId: createDiscipline
      tags: [disciplines]
      summary: Create a discipline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisciplineInput"
      responses:
        "201":
          description: Created discipline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Discipline"
  /disciplines/{disciplineId}:
    patch:
      operationId: updateDiscipline
      tags: [disciplines]
      summary: Update a discipline
      parameters:
        - in: path
          name: disciplineId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisciplineUpdate"
      responses:
        "200":
          description: Updated discipline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Discipline"
    delete:
      operationId: deleteDiscipline
      tags: [disciplines]
      summary: Delete a discipline
      parameters:
        - in: path
          name: disciplineId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── Forms ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, role]
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: ["string", "null"]
        role:
          type: string
        dojo_id:
          type: ["string", "null"]
        stripe_customer_id:
          type: ["string", "null"]
        subscription_status:
          type: string
        camera_side:
          type: ["string", "null"]
        created_at:
          type: string
...
    Student:
      type: object
      required: [id, email]
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: ["string", "null"]
        sessions_count:
          type: integer
        avg_score:
          type: ["number", "null"]
        last_active:
          type: ["string", "null"]

    # ─── Dojo ───────────────────────────────────────────────────────────────
    Dojo:
      type: object
      required: [id, slug, name]
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        logo_url:
          type: ["string", "null"]
        accent_color:
          type: string
        invite_code:
          type: ["string", "null"]
        stripe_customer_id:
          type: ["string", "null"]
        subscription_status:
          type: string
        subscription_tier:
          type: string
        created_at:
          type: string
    DojoPublic:
      type: object
      required: [id, slug, name]
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        logo_url:
          type: ["string", "null"]
        accent_color:
          type: string
        student_count:
          type: integer
        form_count:
          type: integer
    DojoInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        accent_color:
          type: string
        logo_url:
          type: ["string", "null"]
    DojoUpdate:
      type: object
      properties:
        name:
          type: string
        logo_url:
          type: ["string", "null"]
        accent_color:
          type: string
    DojoStats:
      type: object
      required: [student_count, session_count, avg_score, form_count]
      properties:
        student_count:
          type: integer
        session_count:
          type: integer
        avg_score:
          type: ["number", "null"]
        form_count:
          type: integer
    InviteCode:
      type: object
      required: [invite_code]
      properties:
        invite_code:
          type: string
    JoinDojoInput:
      type: object
      required: [invite_code]
      properties:
        invite_code:
          type: string

    # ─── Discipline ─────────────────────────────────────────────────────────
    Discipline:
      type: object
      required: [id, slug, name, is_public]
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        dojo_id:
          type: ["string", "null"]
        is_public:
          type: boolean
        sort_order:
          type: integer
        created_at:
          type: string
    DisciplineInput:
      type: object
      required: [slug, name]
      properties:
        slug:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        dojo_id:
          type: ["string", "null"]
        is_public:
          type: boolean
        sort_order:
          type: integer
    DisciplineUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        is_public:
          type: boolean
        sort_order:
          type: integer

    # ─── Form ───────────────────────────────────────────────────────────────
...