# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FormFlow API - AI Martial Arts Form Trainer
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: string;
  email: string;
  /** @nullable */
  name?: string | null;
  role: string;
  /** @nullable */
  dojo_id?: string | null;
  /** @nullable */
  stripe_customer_id?: string | null;
  subscription_status?: string;
  /** @nullable */
  camera_side?: string | null;
  created_at?: string;
}